requirements.txt
Django>=1.5,<1.6
RapidSMS==0.16.0
South==0.7.6
django-reversion==1.8.0